Move duplicated logic into shared library.
