+++ src/Dfe.SignIn.Web.SelectOrganisation/Program.cs
@@ -1,6 +1,7 @@
 using Azure.Monitor.OpenTelemetry.AspNetCore;
 using Dfe.SignIn.Base.Framework;
 using Dfe.SignIn.Core.Public;
+using Dfe.SignIn.Gateways.SelectOrganisation.DistributedCache;
 using Dfe.SignIn.NodeApi.Client;
 using Dfe.SignIn.Web.SelectOrganisation.Configuration;
 using Dfe.SignIn.WebFramework.Configuration;
@@ -41,7 +42,9 @@ builder.Services
     );
 
 builder.Services
-    .SetupRedisSessionStore(builder.Configuration.GetRequiredSection("SelectOrganisationSessionRedisCache"))
+    .SetupRedisSessionStore(SelectOrganisationConstants.CacheStoreKey,
+        builder.Configuration.GetRequiredSection("SelectOrganisationSessionRedisCache"))
+    .AddSelectOrganisationSessionCache()
     .SetupSelectOrganisationInteractions();
 
 builder.Services
