+++ tests/Dfe.SignIn.Web.SelectOrganisation.UnitTests/Configuration/SelectOrganisationExtensionsTests.cs
@@ -1,9 +1,6 @@
 using Dfe.SignIn.Base.Framework;
 using Dfe.SignIn.Core.Contracts.SelectOrganisation;
-using Dfe.SignIn.Gateways.SelectOrganisation.DistributedCache;
 using Dfe.SignIn.Web.SelectOrganisation.Configuration;
-using Microsoft.Extensions.Caching.Distributed;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Dfe.SignIn.Web.SelectOrganisation.UnitTests.Configuration;
@@ -11,45 +8,6 @@ namespace Dfe.SignIn.Web.SelectOrganisation.UnitTests.Configuration;
 [TestClass]
 public sealed class SelectOrganisationExtensionsTests
 {
-    #region SetupRedisSessionStore(IServiceCollection, IConfiguration)
-
-    [TestMethod]
-    public void SetupRedisSessionStore_Throws_WhenServicesArgumentIsNull()
-    {
-        var configuration = new ConfigurationRoot([]);
-
-        Assert.ThrowsExactly<ArgumentNullException>(()
-            => SelectOrganisationExtensions.SetupRedisSessionStore(null!, configuration));
-    }
-
-    [TestMethod]
-    public void SetupRedisSessionStore_Throws_WhenConfigurationArgumentIsNull()
-    {
-        var services = new ServiceCollection();
-
-        Assert.ThrowsExactly<ArgumentNullException>(()
-            => services.SetupRedisSessionStore(null!));
-    }
-
-    [TestMethod]
-    public void SetupRedisSessionStore_HasExpectedServices()
-    {
-        var configuration = new ConfigurationRoot([]);
-        var services = new ServiceCollection();
-
-        services.SetupRedisSessionStore(configuration);
-
-        Assert.IsTrue(
-            services.Any(descriptor =>
-                (string?)descriptor.ServiceKey == SelectOrganisationConstants.CacheStoreKey &&
-                descriptor.Lifetime == ServiceLifetime.Singleton &&
-                descriptor.ServiceType == typeof(IDistributedCache)
-            )
-        );
-    }
-
-    #endregion
-
     #region SetupSelectOrganisationInteractions(IServiceCollection)
 
     [TestMethod]
